+++ riscv-peripheral/CHANGELOG.md
@@ -11,7 +11,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 - Rework of CLINT peripheral to use methods instead of associated functions.
   This change follows the `svd2rust` pattern, making the ecosystem more consistent.
-- Simplify `clint!` macro using the `Deref` trait.
+- Simplify `clint_codegen!` macro using the `Deref` trait.
+- Rework of PLIC peripherals to use methods instead of associated functions.
+  This change follows the `svd2rust` pattern, making the ecosystem more consistent.
+- Simplify `plic_codegen!` macro using the `Deref` trait.
 
 ### Removed